changed setInterval to timeout on positive thoughts because infinite loops are bad
